+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Added proper Just My Code enabled stacktrace.
 - Added `hitBreakpointIds` support in Stopped Event.
 - Added `hitCondition` support for SourceBreakpoint and FunctionBreakpoint.
+- Added `isOptimized` support in Module.
 
 #### Changed
 - Replaced VSCode to DAP (variables, class names, tests, etc).